chore: satisfy runtime size and test ratchets
