+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/workstation_routes.go
@@ -768,80 +768,6 @@ func resolveRuntimeWorkstationDefinition(
 	return interpolated, found, err
 }
 
-// orderedRuntimeWorkDispatchTokens preserves the authored workstation input
-// order for detached execution. The scheduler intentionally records observed
-// child tokens before the consumed parent token; prompt templates and model
-// operation bindings, however, address inputs by the workstation's declared
-// slots. The legacy workstation executor made this ordering adjustment before
-// it built its request, so the shared Workers execution path must do the same.
-func orderedRuntimeWorkDispatchTokens(
-	cfg *runtimeConfig,
-	request workers.WorkstationDispatchRequest,
-	invocation *work.InvocationArguments,
-) ([]workers.Token, error) {
-	tokens := workers.WorkDispatchInputTokens(request.Execution.Dispatch)
-	if len(tokens) < 2 {
-		return tokens, nil
-	}
-	lookup, ok := runtimeDefinitionLookup(cfg)
-	if !ok {
-		return tokens, nil
-	}
-	workstation, found, err := resolveRuntimeWorkstationDefinition(
-		cfg,
-		lookup,
-		request,
-		invocation,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("resolve workstation input order: %w", err)
-	}
-	if !found || workstation == nil {
-		return tokens, nil
-	}
-
-	byPlace := make(map[string][]int)
-	for index, token := range tokens {
-		byPlace[workerTokenPlaceKey(token)] = append(byPlace[workerTokenPlaceKey(token)], index)
-	}
-	ordered := make([]workers.Token, 0, len(tokens))
-	used := make([]bool, len(tokens))
-	appendPlaceTokens := func(placeID string) {
-		for _, index := range byPlace[placeID] {
-			used[index] = true
-			ordered = append(ordered, tokens[index])
-		}
-	}
-	for _, input := range workstation.Inputs {
-		appendPlaceTokens(fmt.Sprintf("%s:%s", input.WorkTypeName, input.StateName))
-	}
-	for _, resource := range workstation.Resources {
-		appendPlaceTokens(fmt.Sprintf("%s:%s", resource.Name, interfaces.ResourceStateAvailable))
-	}
-	for index, token := range tokens {
-		if used[index] {
-			continue
-		}
-		ordered = append(ordered, token)
-	}
-	return ordered, nil
-}
-
-func workerTokenPlaceKey(token workers.Token) string {
-	prefix := strings.TrimSpace(token.Color.WorkTypeID)
-	if prefix == "" {
-		prefix = strings.TrimSpace(token.Color.Name)
-	}
-	stateName := strings.TrimSpace(token.State)
-	if prefix == "" {
-		return stateName
-	}
-	if stateName == "" {
-		return prefix
-	}
-	return prefix + ":" + stateName
-}
-
 func resolveRuntimeWorkerDefinition(
 	lookup interfaces.RuntimeDefinitionLookup,
 	selection *runtimeExecutionSelection,
